Fix infinity consumer close method
Before this patch, if there is no connection to the broker,
then the close method of consumer could hang while clearing
the message queue due to missing error handling.

Now, if it is impossible to get messages from the queue for a long time,
then the queue stops clearing and the close method ends its work.