patch: fix deref after NULL in rdkafka_conf
In `rd_kafka_confval_set_type` at `RD_KAFKA_CONFVAL_STR` case
variable `v` is checked for NULL, after which it is dereferenced
uncoditionally by `strlen` and `rd_strdup`.

Let's set it's length to 0, if `v` equals to NULL. Error will be
throws below, if it's forbidden. If it's not, we assign dynamically
allocated empty string to a value instead of calling `rd_strdup`
on NULL.

Closes tarantool/security#52