patch: fix deref after NULL in lz4
If `dictBase` variable equals to NULL, we may reference NULL while
reading from `match`. Even thought the problem was tried to be solved
in lz4/lz4@59273b7, it won't help us, as the commit adds nothing but
assertions and we don't build librdkafka with DEVEL option.

So, let's check if `match` is not NULL right before reading from it.

Part of tarantool/security#70