bump librdkafka to tag 'v2.7.0'/'v2.6.1'
The problem is that the tag in a previous version had a bug:
(https://github.com/confluentinc/librdkafka/pull/4895).
And it was fixed in this version.

We can't use version 2.8.0 for release since it's a bit incompatible
with a current SDK (because it's require some symbols).
So plan is following:
  - Downgrade librdkafka version to old version.
  - Release new version
  - Revert this commit.